+++ packages/onramp-kit/example/client/.env.sample
@@ -6,4 +6,8 @@ VITE_STRIPE_PUBLIC_KEY=
 # More info here:
 #   https://docs.safe.global/learn/safe-core/safe-core-account-abstraction-sdk/onramp-kit#prerequisites
 VITE_SAFE_STRIPE_BACKEND_BASE_URL=
+
+# Configure the Monerium client ID. You need to get one from Monerium.
+# More info here:
+#   https://monerium.dev/docs/getting-started/create-app
 VITE_MONERIUM_CLIENT_ID=
